Keep process integration tests out of the fast push checks
